基本題對了！

+++ hw4/OS_homework4_bonus/main.cu
@@ -51,8 +51,8 @@ __device__ __managed__ uchar *volume;
 int load_binaryFile(const char *filename, uchar *a, int max_size);
 void write_binaryFIle(const char *filename, uchar *a, int size);
 __device__ u32 open(const char *name, uchar mode);
-__device__ uchar write(uchar *input, int n, uchar fp);
-__device__ uchar read(uchar *output, int n, uchar fp);
+__device__ uchar write(uchar *input, int n, u32 fp);
+__device__ uchar read(uchar *output, int n, u32 fp);
 __device__ void gsys(uchar arg, const char* file=NULL);
 __device__ void init_volume();
 
@@ -263,7 +263,7 @@ __device__ u32 open(const char *file, uchar mode) {
 	else return ERROR;
 }
 
-__device__ uchar write(uchar *input, int n, uchar fp) {
+__device__ uchar write(uchar *input, int n, u32 fp) {
 	if(fp == ERROR) return WRITE_ERROR;
 	if(n > BLOCK_SIZE) return WRITE_ERROR;
 	int curb = BLOCK_LOC(fp);
@@ -275,7 +275,7 @@ __device__ uchar write(uchar *input, int n, uchar fp) {
 	}
 	return WRITE_SUCCESS;
 }
-__device__ uchar read(uchar *output, int n, uchar fp) {
+__device__ uchar read(uchar *output, int n, u32 fp) {
 	if(fp == ERROR) return READ_ERROR;
 	if(n > BLOCK_SIZE) return READ_ERROR;
 	int curb = BLOCK_LOC(fp);
